@@ -13,66 +13,78 @@
   "scripts": {
     "build": "rm -rf build; webpack",
     "test": "jest",
+    "test:update": "jest -u",
     "test:ci": "jest --runInBand",
     "test:watch": "jest --watch",
-    "test:coverage": "jest --coverage",
     "postinstall": "opencollective-postinstall",
-    "lint": "eslint src/",
-    "precommit": "lint-staged && npm test",
-    "prettify": "prettier --write 'src/**/*.js'"
+    "lint": "eslint src",
+    "prettify": "prettier --single-quote --trailing-comma=es5 --write './**/*.md'",
+    "clean-install": "rm -rf node_modules && yarn"
   },
   "author": "Nader Dabit & Monte Thakkar",
   "license": "MIT",
   "bugs": {
     "url": "https://github.com/react-native-training/react-native-elements/issues"
   },
-  "homepage": "https://github.com/react-native-training/react-native-elements#readme",
+  "homepage": "https://react-native-training.github.io/react-native-elements/",
   "collective": {
     "type": "opencollective",
     "url": "https://opencollective.com/react-native-elements",
     "logo": "https://opencollective.com/react-native-elements/logo.txt"
   },
   "dependencies": {
+    "color": "^3.1.0",
+    "hoist-non-react-statics": "^3.1.0",
     "lodash.merge": "^4.6.1",
-    "lodash.times": "^4.3.2",
     "opencollective-postinstall": "^2.0.0",
-    "prop-types": "^15.5.8"
+    "prop-types": "^15.5.8",
+    "react-native-ratings": "^6.3.0",
+    "react-native-status-bar-height": "^2.2.0"
   },
   "devDependencies": {
     "@types/react": "^16.4.16",
     "@types/react-native": "^0.57.4",
     "babel": "^6.23.0",
     "babel-core": "^6.23.1",
-    "babel-eslint": "^7.1.1",
+    "babel-eslint": "^10.0.1",
     "babel-jest": "22.4.3",
     "babel-loader": "^6.3.2",
     "babel-preset-react-native": "4.0.0",
-    "codecov": "^2.1.0",
     "enzyme": "^3.7.0",
     "enzyme-adapter-react-16": "^1.6.0",
     "enzyme-to-json": "^3.3.4",
-    "eslint": "^3.17.1",
-    "eslint-plugin-react": "^6.10.0",
-    "eslint-plugin-react-native": "^2.3.1",
+    "eslint": "^5.10.0",
+    "eslint-config-airbnb": "^17.1.0",
+    "eslint-config-prettier": "^3.3.0",
+    "eslint-plugin-import": "^2.14.0",
+    "eslint-plugin-jest": "^22.1.2",
+    "eslint-plugin-jsx-a11y": "^6.1.2",
+    "eslint-plugin-prettier": "^3.0.0",
+    "eslint-plugin-react": "^7.11.1",
+    "eslint-plugin-react-native": "^3.5.0",
     "file-loader": "^0.11.1",
-    "husky": "^0.13.3",
+    "husky": "^1.1.2",
     "image-webpack-loader": "^3.3.1",
     "jest": "23.5.0",
-    "lint-staged": "^3.4.0",
-    "prettier": "^1.10.2",
+    "jest-transform-stub": "^1.0.0",
+    "lint-staged": "^8.0.4",
+    "prettier": "^1.15.3",
     "react": "16.3.1",
     "react-dom": "16.2.0",
     "react-native": "0.55.4",
-    "react-native-vector-icons": "^4.4.2",
-    "react-test-renderer": "16.3.1",
+    "react-native-vector-icons": "^6.0.2",
+    "react-test-renderer": "^16.6.0",
     "webpack": "^2.2.1"
   },
   "peerDependencies": {
-    "react-native-vector-icons": "^4.2.0 || ^5.0.0"
+    "react-native-vector-icons": ">4.2.0"
   },
   "jest": {
     "preset": "react-native",
     "coverageDirectory": "./coverage/",
+    "testPathIgnorePatterns": [
+      "./src/searchbar/__tests__/common.js"
+    ],
     "collectCoverageFrom": [
       "src/**/*.js",
       "!src/index.js",
