@@ -9,8 +9,7 @@
 // version: 03
 //
 
-const React = require('react-native'); // eslint-disable-line no-undef
-const { PixelRatio, Dimensions } = React;
+import { PixelRatio, Dimensions } from 'react-native';
 
 const pixelRatio = PixelRatio.get();
 const deviceHeight = Dimensions.get('window').height;
@@ -31,53 +30,66 @@ const normalize = size => {
     if (deviceWidth < 360) {
       return size * 0.95;
     }
+
     // iphone 5
     if (deviceHeight < 667) {
       return size;
       // iphone 6-6s
-    } else if (deviceHeight >= 667 && deviceHeight <= 735) {
+    }
+
+    if (deviceHeight >= 667 && deviceHeight <= 735) {
       return size * 1.15;
     }
     // older phablets
     return size * 1.25;
-  } else if (pixelRatio >= 3 && pixelRatio < 3.5) {
+  }
+
+  if (pixelRatio >= 3 && pixelRatio < 3.5) {
     // catch Android font scaling on small machines
     // where pixel ratio / font scale ratio => 3:3
     if (deviceWidth <= 360) {
       return size;
     }
+
     // Catch other weird android width sizings
     if (deviceHeight < 667) {
       return size * 1.15;
       // catch in-between size Androids and scale font up
       // a tad but not too much
     }
+
     if (deviceHeight >= 667 && deviceHeight <= 735) {
       return size * 1.2;
     }
+
     // catch larger devices
     // ie iphone 6s plus / 7 plus / mi note 等等
     return size * 1.27;
-  } else if (pixelRatio >= 3.5) {
+  }
+
+  if (pixelRatio >= 3.5) {
     // catch Android font scaling on small machines
     // where pixel ratio / font scale ratio => 3:3
     if (deviceWidth <= 360) {
       return size;
       // Catch other smaller android height sizings
     }
+
     if (deviceHeight < 667) {
       return size * 1.2;
       // catch in-between size Androids and scale font up
       // a tad but not too much
     }
+
     if (deviceHeight >= 667 && deviceHeight <= 735) {
       return size * 1.25;
     }
+
     // catch larger phablet devices
     return size * 1.4;
-  } else
-    // if older device ie pixelRatio !== 2 || 3 || 3.5
-    return size;
+  }
+
+  return size;
 };
 
 module.exports = normalize; // eslint-disable-line no-undef
