@@ -3,33 +3,39 @@ import PropTypes from 'prop-types';
 import {
   View,
   Text,
-  Image,
+  Image as RNImage,
   Platform,
   StyleSheet,
   TouchableOpacity,
   TouchableHighlight,
   TouchableNativeFeedback,
   TouchableWithoutFeedback,
-  Animated,
 } from 'react-native';
 
 import { withTheme, ViewPropTypes } from '../config';
 import { renderNode, nodeType } from '../helpers';
 
 import Icon from '../icons/Icon';
+import Image from '../image/Image';
 
-const DEFAULT_COLORS = ['#000', '#333', '#555', '#888', '#aaa', '#ddd'];
-const DEFAULT_SIZES = {
+const avatarSizes = {
   small: 34,
   medium: 50,
   large: 75,
   xlarge: 150,
 };
 
+const defaultEditButton = {
+  name: 'mode-edit',
+  type: 'material',
+  color: '#fff',
+  underlayColor: '#000',
+};
+
 const Avatar = ({
   onPress,
   onLongPress,
-  component: Component = onPress || onLongPress ? TouchableOpacity : View,
+  Component = onPress || onLongPress ? TouchableOpacity : View,
   containerStyle,
   icon,
   iconStyle,
@@ -41,7 +47,7 @@ const Avatar = ({
   titleStyle,
   overlayContainerStyle,
   showEditButton,
-  editButton,
+  editButton: passedEditButton,
   onEditPress,
   imageProps,
   placeholderStyle,
@@ -50,17 +56,20 @@ const Avatar = ({
   ...attributes
 }) => {
   const width =
-    typeof size === 'number'
-      ? size
-      : DEFAULT_SIZES[size] || DEFAULT_SIZES.small;
+    typeof size === 'number' ? size : avatarSizes[size] || avatarSizes.small;
   const height = width;
   const titleSize = width / 2;
   const iconSize = width / 2;
+
+  const editButton = {
+    ...defaultEditButton,
+    ...passedEditButton,
+  };
   const editButtonSize = editButton.size || (width + height) / 2 / 3;
 
   const Utils = showEditButton && (
     <TouchableHighlight
-      style={[
+      style={StyleSheet.flatten([
         styles.editButton,
         {
           width: editButtonSize,
@@ -68,17 +77,12 @@ const Avatar = ({
           borderRadius: editButtonSize / 2,
         },
         editButton.style,
-      ]}
+      ])}
       underlayColor={editButton.underlayColor}
       onPress={onEditPress}
     >
       <View>
-        <Icon
-          size={editButtonSize * 0.8}
-          name={editButton.iconName}
-          type={editButton.iconType}
-          color={editButton.iconColor}
-        />
+        <Icon size={editButtonSize * 0.8} {...editButton} />
       </View>
     </TouchableHighlight>
   );
@@ -87,7 +91,13 @@ const Avatar = ({
     (renderPlaceholderContent &&
       renderNode(undefined, renderPlaceholderContent)) ||
     (title && (
-      <Text style={[styles.title, { fontSize: titleSize }, titleStyle]}>
+      <Text
+        style={StyleSheet.flatten([
+          styles.title,
+          { fontSize: titleSize },
+          titleStyle,
+        ])}
+      >
         {title}
       </Text>
     )) ||
@@ -101,25 +111,39 @@ const Avatar = ({
       />
     ));
 
+  // Remove placeholder styling if we're not using image
+  const hidePlaceholder = !source;
+
   return (
     <Component
       onPress={onPress}
       onLongPress={onLongPress}
-      style={[
+      style={StyleSheet.flatten([
         styles.container,
         { height, width },
-        rounded && { borderRadius: width / 2, overflow: 'hidden' },
+        rounded && { borderRadius: width / 2 },
         containerStyle,
-      ]}
+      ])}
       {...attributes}
     >
-      <FadeInImage
-        placeholderStyle={placeholderStyle}
+      <Image
+        placeholderStyle={StyleSheet.flatten([
+          placeholderStyle,
+          hidePlaceholder && { backgroundColor: 'transparent' },
+        ])}
         PlaceholderContent={PlaceholderContent}
-        containerStyle={overlayContainerStyle}
+        containerStyle={StyleSheet.flatten([
+          styles.overlayContainer,
+          overlayContainerStyle,
+          rounded && { borderRadius: width / 2, overflow: 'hidden' },
+        ])}
         source={source}
         {...imageProps}
-        style={[imageProps && imageProps.style, avatarStyle]}
+        style={StyleSheet.flatten([
+          styles.avatar,
+          imageProps && imageProps.style,
+          avatarStyle,
+        ])}
         ImageComponent={ImageComponent}
       />
       {Utils}
@@ -138,6 +162,7 @@ const styles = StyleSheet.create({
   },
   overlayContainer: {
     flex: 1,
+    backgroundColor: '#bdbdbd',
   },
   title: {
     color: '#ffffff',
@@ -150,32 +175,23 @@ const styles = StyleSheet.create({
     right: 0,
     alignItems: 'center',
     justifyContent: 'center',
-    backgroundColor: DEFAULT_COLORS[4],
+    backgroundColor: '#aaa',
     ...Platform.select({
-      ios: {
-        shadowColor: DEFAULT_COLORS[0],
+      android: {
+        elevation: 1,
+      },
+      default: {
+        shadowColor: '#000',
         shadowOffset: { width: 1, height: 1 },
         shadowRadius: 2,
         shadowOpacity: 0.5,
       },
-      android: {
-        elevation: 1,
-      },
     }),
-  },
-  placeholderContainer: {
-    ...StyleSheet.absoluteFillObject,
-  },
-  placeholder: {
-    flex: 1,
-    alignItems: 'center',
-    justifyContent: 'center',
-    backgroundColor: '#BDBDBD',
   },
 });
 
 Avatar.propTypes = {
-  component: PropTypes.oneOf([
+  Component: PropTypes.oneOf([
     View,
     TouchableOpacity,
     TouchableHighlight,
@@ -184,13 +200,13 @@ Avatar.propTypes = {
   ]),
   onPress: PropTypes.func,
   onLongPress: PropTypes.func,
-  containerStyle: PropTypes.any,
-  source: Image.propTypes.source,
-  avatarStyle: PropTypes.any,
+  containerStyle: ViewPropTypes.style,
+  source: RNImage.propTypes.source,
+  avatarStyle: ViewPropTypes.style,
   rounded: PropTypes.bool,
   title: PropTypes.string,
   titleStyle: Text.propTypes.style,
-  overlayContainerStyle: PropTypes.any,
+  overlayContainerStyle: ViewPropTypes.style,
   activeOpacity: PropTypes.number,
   icon: PropTypes.object,
   iconStyle: Text.propTypes.style,
@@ -202,91 +218,25 @@ Avatar.propTypes = {
   onEditPress: PropTypes.func,
   editButton: PropTypes.shape({
     size: PropTypes.number,
-    iconName: PropTypes.string,
-    iconType: PropTypes.string,
-    iconColor: PropTypes.string,
+    name: PropTypes.string,
+    type: PropTypes.string,
+    color: PropTypes.string,
     underlayColor: PropTypes.string,
     style: ViewPropTypes.style,
   }),
   placeholderStyle: ViewPropTypes.style,
   renderPlaceholderContent: nodeType,
   imageProps: PropTypes.object,
-  ImageComponent: PropTypes.func,
+  ImageComponent: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
 };
 
 Avatar.defaultProps = {
   showEditButton: false,
   onEditPress: null,
   size: 'small',
-  editButton: {
-    size: null,
-    iconName: 'mode-edit',
-    iconType: 'material',
-    iconColor: '#fff',
-    underlayColor: DEFAULT_COLORS[0],
-    style: null,
-  },
-  ImageComponent: Image,
+  editButton: defaultEditButton,
+  ImageComponent: RNImage,
 };
 
-class FadeInImage extends React.PureComponent {
-  placeholderContainerOpacity = new Animated.Value(1);
-
-  onLoadEnd = () => {
-    /* Images finish loading in the same frame for some reason,
-      the images will fade in separately with staggerNonce */
-    const minimumWait = 100;
-    const staggerNonce = 200 * Math.random();
-    setTimeout(
-      () =>
-        Animated.timing(this.placeholderContainerOpacity, {
-          toValue: 0,
-          duration: 350,
-          useNativeDriver: true,
-        }).start(),
-      minimumWait + staggerNonce
-    );
-  };
-
-  render() {
-    const {
-      placeholderStyle,
-      PlaceholderContent,
-      containerStyle,
-      style,
-      ImageComponent,
-      ...attributes
-    } = this.props;
-
-    return Platform.OS === 'ios' ? (
-      <View style={[styles.overlayContainer, containerStyle]}>
-        <ImageComponent
-          {...attributes}
-          onLoadEnd={this.onLoadEnd}
-          style={[styles.avatar, style]}
-        />
-        <Animated.View
-          style={[
-            styles.placeholderContainer,
-            { opacity: this.placeholderContainerOpacity },
-          ]}
-        >
-          <View style={[style, styles.placeholder, placeholderStyle]}>
-            {PlaceholderContent}
-          </View>
-        </Animated.View>
-      </View>
-    ) : (
-      <View style={[styles.overlayContainer, containerStyle]}>
-        <View style={styles.placeholderContainer}>
-          <View style={[style, styles.placeholder, placeholderStyle]}>
-            {PlaceholderContent}
-          </View>
-        </View>
-        <Image {...attributes} style={[styles.avatar, style]} />
-      </View>
-    );
-  }
-}
-
+export { Avatar };
 export default withTheme(Avatar, 'Avatar');
