@@ -1,3 +1,4 @@
+/* eslint-disable react/no-array-index-key */
 import React from 'react';
 import PropTypes from 'prop-types';
 import {
@@ -10,14 +11,15 @@ import {
 } from 'react-native';
 
 import { ViewPropTypes, withTheme } from '../config';
+import { normalizeText, color } from '../helpers';
+
 import Text from '../text/Text';
-import normalize from '../helpers/normalizeText';
 
 const ButtonGroup = props => {
   const { theme, ...rest } = props;
 
   const {
-    component: Component,
+    Component,
     buttons,
     onPress,
     selectedIndex,
@@ -36,13 +38,20 @@ const ButtonGroup = props => {
     onShowUnderlay,
     setOpacityTo,
     containerBorderRadius,
-    disableSelected,
+    disabled,
+    disabledStyle,
+    disabledTextStyle,
+    disabledSelectedStyle,
+    disabledSelectedTextStyle,
     ...attributes
   } = rest;
 
   let innerBorderWidth = 1;
 
-  if (innerBorderStyle && innerBorderStyle.hasOwnProperty('width')) {
+  if (
+    innerBorderStyle &&
+    Object.prototype.hasOwnProperty.call(innerBorderStyle, 'width')
+  ) {
     innerBorderWidth = innerBorderStyle.width;
   }
 
@@ -56,6 +65,9 @@ const ButtonGroup = props => {
     >
       {buttons.map((button, i) => {
         const isSelected = selectedIndex === i || selectedIndexes.includes(i);
+        const isDisabled =
+          disabled === true ||
+          (Array.isArray(disabled) && disabled.includes(i));
 
         return (
           <View
@@ -88,12 +100,13 @@ const ButtonGroup = props => {
             ])}
           >
             <Component
+              testID="buttonGroupItem"
               activeOpacity={activeOpacity}
               setOpacityTo={setOpacityTo}
               onHideUnderlay={onHideUnderlay}
               onShowUnderlay={onShowUnderlay}
               underlayColor={underlayColor}
-              disabled={disableSelected && isSelected ? true : false}
+              disabled={isDisabled}
               onPress={() => {
                 if (selectMultiple) {
                   if (selectedIndexes.includes(i)) {
@@ -115,6 +128,10 @@ const ButtonGroup = props => {
                     backgroundColor: theme.colors.primary,
                   },
                   isSelected && selectedButtonStyle && selectedButtonStyle,
+                  isDisabled && styles.disabled,
+                  isDisabled && disabledStyle,
+                  isDisabled && isSelected && styles.disabledSelected(theme),
+                  isDisabled && isSelected && disabledSelectedStyle,
                 ])}
               >
                 {button.element ? (
@@ -127,6 +144,9 @@ const ButtonGroup = props => {
                       textStyle && textStyle,
                       isSelected && { color: '#fff' },
                       isSelected && selectedTextStyle,
+                      isDisabled && styles.disabledText(theme),
+                      isDisabled && disabledTextStyle,
+                      isDisabled && isSelected && disabledSelectedTextStyle,
                     ])}
                   >
                     {button}
@@ -164,19 +184,29 @@ const styles = {
     height: 40,
   },
   buttonText: theme => ({
-    fontSize: normalize(13),
+    fontSize: normalizeText(13),
     color: theme.colors.grey2,
     ...Platform.select({
-      ios: {
+      android: {},
+      default: {
         fontWeight: '500',
       },
     }),
+  }),
+  disabled: {
+    backgroundColor: 'transparent',
+  },
+  disabledText: theme => ({
+    color: color(theme.colors.disabled).darken(0.3),
+  }),
+  disabledSelected: theme => ({
+    backgroundColor: theme.colors.disabled,
   }),
 };
 
 ButtonGroup.propTypes = {
   button: PropTypes.object,
-  component: PropTypes.any,
+  Component: PropTypes.oneOfType([PropTypes.func, PropTypes.object]),
   onPress: PropTypes.func,
   buttons: PropTypes.array,
   containerStyle: ViewPropTypes.style,
@@ -189,7 +219,7 @@ ButtonGroup.propTypes = {
   activeOpacity: PropTypes.number,
   onHideUnderlay: PropTypes.func,
   onShowUnderlay: PropTypes.func,
-  setOpacityTo: PropTypes.any,
+  setOpacityTo: PropTypes.func,
   innerBorderStyle: PropTypes.shape({
     color: PropTypes.string,
     width: PropTypes.number,
@@ -200,17 +230,28 @@ ButtonGroup.propTypes = {
   ]),
   buttonStyle: ViewPropTypes.style,
   containerBorderRadius: PropTypes.number,
-  disableSelected: PropTypes.bool,
   selectMultiple: PropTypes.bool,
   theme: PropTypes.object,
+  disabled: PropTypes.oneOfType([
+    PropTypes.bool,
+    PropTypes.arrayOf(PropTypes.number),
+  ]),
+  disabledStyle: ViewPropTypes.style,
+  disabledTextStyle: NativeText.propTypes.style,
+  disabledSelectedStyle: ViewPropTypes.style,
+  disabledSelectedTextStyle: NativeText.propTypes.style,
 };
 
 ButtonGroup.defaultProps = {
   selectedIndexes: [],
   selectMultiple: false,
   containerBorderRadius: 3,
-  onPress: () => {},
-  component: Platform.OS === 'ios' ? TouchableOpacity : TouchableNativeFeedback,
+  disabled: false,
+  Component: Platform.select({
+    android: TouchableNativeFeedback,
+    default: TouchableOpacity,
+  }),
+  onPress: () => null,
 };
 
 export { ButtonGroup };
