@@ -41,14 +41,14 @@ describe('Button Component', () => {
       .props()
       .onPress();
 
-    expect(onPress).toBeCalled();
+    expect(onPress).toHaveBeenCalled();
   });
 
   it('should have ripple on android version 21 and higher', () => {
     jest.mock('Platform', () => ({
       OS: 'android',
       Version: 25,
-      select: function() {},
+      select() {},
     }));
 
     const wrapper = shallow(<Button theme={theme} />);
@@ -60,7 +60,7 @@ describe('Button Component', () => {
     jest.mock('Platform', () => ({
       OS: 'android',
       Version: 20,
-      select: function() {},
+      select() {},
     }));
 
     const wrapper = shallow(<Button theme={theme} />);
@@ -82,15 +82,93 @@ describe('Button Component', () => {
     );
   });
 
+  describe('Button Types', () => {
+    describe('Solid', () => {
+      it('should display solid button', () => {
+        const component = shallow(<Button theme={theme} title="Solid" />);
+        expect(component.length).toBe(1);
+        expect(toJson(component)).toMatchSnapshot();
+      });
+
+      it('should display raised solid button', () => {
+        const component = shallow(
+          <Button theme={theme} title="Solid" raised />
+        );
+        expect(component.length).toBe(1);
+        expect(toJson(component)).toMatchSnapshot();
+      });
+
+      it('should display solid button disabled', () => {
+        const component = shallow(
+          <Button theme={theme} title="Solid" disabled />
+        );
+        expect(component.length).toBe(1);
+        expect(toJson(component)).toMatchSnapshot();
+      });
+    });
+
+    describe('Outline', () => {
+      it('should display outline button', () => {
+        const component = shallow(
+          <Button theme={theme} title="Outline" type="outline" />
+        );
+        expect(component.length).toBe(1);
+        expect(toJson(component)).toMatchSnapshot();
+      });
+
+      it('should display raised outline button', () => {
+        const component = shallow(
+          <Button theme={theme} title="Outline" type="outline" raised />
+        );
+        expect(component.length).toBe(1);
+        expect(toJson(component)).toMatchSnapshot();
+      });
+
+      it('should display outline button disabled', () => {
+        const component = shallow(
+          <Button theme={theme} title="Outline" type="outline" disabled />
+        );
+        expect(component.length).toBe(1);
+        expect(toJson(component)).toMatchSnapshot();
+      });
+    });
+
+    describe('Clear', () => {
+      it('should display clear button', () => {
+        const component = shallow(
+          <Button theme={theme} title="Clear" type="clear" />
+        );
+        expect(component.length).toBe(1);
+        expect(toJson(component)).toMatchSnapshot();
+      });
+
+      it('should display raised clear button', () => {
+        const component = shallow(
+          <Button theme={theme} title="Clear" type="clear" raised />
+        );
+        expect(component.length).toBe(1);
+        expect(toJson(component)).toMatchSnapshot();
+      });
+
+      it('should display clear button disabled', () => {
+        const component = shallow(
+          <Button theme={theme} title="Clear" type="clear" disabled />
+        );
+        expect(component.length).toBe(1);
+        expect(toJson(component)).toMatchSnapshot();
+      });
+    });
+  });
+
   it('should apply values from theme', () => {
-    const theme = {
+    const testTheme = {
       Button: {
         loading: true,
       },
     };
 
     const component = create(
-      <ThemeProvider theme={theme}>
+      <ThemeProvider theme={testTheme}>
         <ThemedButton />
       </ThemeProvider>
     );
