@@ -17,7 +17,6 @@ import Divider from '../divider/Divider';
 const Card = props => {
   const {
     children,
-    flexDirection,
     containerStyle,
     wrapperStyle,
     imageWrapperStyle,
@@ -31,7 +30,6 @@ const Card = props => {
     dividerStyle,
     image,
     imageStyle,
-    fontFamily,
     imageProps,
     theme,
     ...attributes
@@ -50,7 +48,6 @@ const Card = props => {
         style={StyleSheet.flatten([
           styles.wrapper,
           wrapperStyle && wrapperStyle,
-          flexDirection && { flexDirection },
         ])}
       >
         {title === '' || React.isValidElement(title)
@@ -64,12 +61,12 @@ const Card = props => {
                     styles.cardTitle(theme),
                     image && styles.imageCardTitle,
                     titleStyle && titleStyle,
-                    fontFamily && { fontFamily },
                   ])}
                   numberOfLines={titleNumberOfLines}
                 >
                   {title}
                 </Text>
+
                 {!image && (
                   <Divider
                     style={StyleSheet.flatten([
@@ -80,6 +77,7 @@ const Card = props => {
                 )}
               </View>
             )}
+
         {image && (
           <View style={imageWrapperStyle && imageWrapperStyle}>
             <BackgroundImage
@@ -112,6 +110,7 @@ const Card = props => {
                 </View>
               )}
             </BackgroundImage>
+
             <View
               style={StyleSheet.flatten([
                 { padding: 10 },
@@ -122,6 +121,7 @@ const Card = props => {
             </View>
           </View>
         )}
+
         {!image && children}
       </View>
     </View>
@@ -129,8 +129,10 @@ const Card = props => {
 };
 
 Card.propTypes = {
-  children: PropTypes.any,
-  flexDirection: PropTypes.string,
+  children: PropTypes.oneOfType([
+    PropTypes.element,
+    PropTypes.arrayOf(PropTypes.element),
+  ]),
   containerStyle: ViewPropTypes.style,
   wrapperStyle: ViewPropTypes.style,
   overlayStyle: ViewPropTypes.style,
@@ -144,7 +146,6 @@ Card.propTypes = {
   image: Image.propTypes.source,
   imageStyle: ViewPropTypes.style,
   imageWrapperStyle: ViewPropTypes.style,
-  fontFamily: PropTypes.string,
   imageProps: PropTypes.object,
   titleNumberOfLines: PropTypes.number,
   theme: PropTypes.object,
@@ -159,14 +160,14 @@ const styles = {
     marginBottom: 0,
     borderColor: theme.colors.grey5,
     ...Platform.select({
-      ios: {
+      android: {
+        elevation: 1,
+      },
+      default: {
         shadowColor: 'rgba(0,0,0, .2)',
         shadowOffset: { height: 0, width: 0 },
         shadowOpacity: 1,
         shadowRadius: 1,
-      },
-      android: {
-        elevation: 1,
       },
     }),
   }),
@@ -175,11 +176,11 @@ const styles = {
     marginBottom: 8,
     color: 'white',
     ...Platform.select({
-      ios: {
-        fontWeight: '800',
-      },
       android: {
         ...fonts.android.black,
+      },
+      default: {
+        fontWeight: '800',
       },
     }),
   },
@@ -188,11 +189,11 @@ const styles = {
     marginBottom: 8,
     color: 'white',
     ...Platform.select({
-      ios: {
-        fontWeight: '400',
-      },
       android: {
         ...fonts.android.black,
+      },
+      default: {
+        fontWeight: '400',
       },
     }),
   },
@@ -206,11 +207,11 @@ const styles = {
     fontSize: normalize(14),
     color: theme.colors.grey1,
     ...Platform.select({
-      ios: {
-        fontWeight: 'bold',
-      },
       android: {
         ...fonts.android.black,
+      },
+      default: {
+        fontWeight: 'bold',
       },
     }),
     textAlign: 'center',
