@@ -22,13 +22,11 @@ export default class ThemeProvider extends React.Component {
 
   updateTheme = updates => {
     this.setState(({ theme }) => ({
-      theme: merge(theme, updates),
+      theme: merge({}, theme, updates),
     }));
   };
 
-  getTheme = () => {
-    return this.state.theme;
-  };
+  getTheme = () => this.state.theme;
 
   render() {
     return (
@@ -47,6 +45,10 @@ export default class ThemeProvider extends React.Component {
 ThemeProvider.propTypes = {
   theme: PropTypes.object,
   children: PropTypes.node.isRequired,
+};
+
+ThemeProvider.defaultProps = {
+  theme: {},
 };
 
 export const ThemeConsumer = ThemeContext.Consumer;
