@@ -1,11 +1,11 @@
 import { Text } from 'react-native';
+import { getStatusBarHeight } from 'react-native-status-bar-height';
+
 import BackgroundImage from './BackgroundImage';
 import colors from './colors';
-import { getStatusBarHeight } from './statusBar';
 import ViewPropTypes from './ViewPropTypes';
 import fonts from './fonts';
 import ThemeProvider, { ThemeConsumer } from './ThemeProvider';
-import merge from 'lodash.merge';
 import withTheme from './withTheme';
 
 const TextPropTypes = Text.propTypes;
@@ -19,6 +19,5 @@ export {
   fonts,
   ThemeProvider,
   ThemeConsumer,
-  merge,
   withTheme,
 };
