@@ -9,6 +9,7 @@ import EntypoIcon from 'react-native-vector-icons/Entypo';
 import FAIcon from 'react-native-vector-icons/FontAwesome';
 import SimpleLineIcon from 'react-native-vector-icons/SimpleLineIcons';
 import FeatherIcon from 'react-native-vector-icons/Feather';
+import AntIcon from 'react-native-vector-icons/AntDesign';
 
 const customIcons = {};
 
@@ -40,8 +41,10 @@ export default type => {
       return SimpleLineIcon;
     case 'feather':
       return FeatherIcon;
+    case 'antdesign':
+      return AntIcon;
     default:
-      if (customIcons.hasOwnProperty(type)) {
+      if (Object.prototype.hasOwnProperty.call(customIcons, type)) {
         return customIcons[type];
       }
       return MaterialIcon;
