@@ -1,6 +1,6 @@
 import React from 'react';
 
-export default (Component, content, defaultProps) => {
+const renderNode = (Component, content, defaultProps) => {
   if (content == null || content === false) {
     return null;
   }
@@ -14,8 +14,10 @@ export default (Component, content, defaultProps) => {
   if (content === true) {
     return <Component {...defaultProps} />;
   }
-  if (typeof content === 'string') {
+  if (typeof content === 'string' || typeof content === 'number') {
     return <Component {...defaultProps}>{content}</Component>;
   }
   return <Component {...defaultProps} {...content} />;
 };
+
+export default renderNode;
