@@ -26,11 +26,18 @@ const Icon = props => {
     disabled,
     disabledStyle,
     onPress,
-    component: Component = onPress ? TouchableHighlight : View,
+    Component = onPress ? TouchableHighlight : View,
     ...attributes
   } = props;
 
-  let Icon = getIconType(type || 'material');
+  const IconComponent = getIconType(type);
+  const getBackgroundColor = () => {
+    if (reverse) {
+      return color;
+    }
+
+    return raised ? 'white' : 'transparent';
+  };
 
   return (
     <View style={containerStyle && containerStyle}>
@@ -46,7 +53,7 @@ const Icon = props => {
           },
           raised && styles.raised,
           {
-            backgroundColor: reverse ? color : raised ? 'white' : 'transparent',
+            backgroundColor: getBackgroundColor(),
             alignItems: 'center',
             justifyContent: 'center',
           },
@@ -56,7 +63,7 @@ const Icon = props => {
         {...onPress && { disabled }}
         onPress={onPress}
       >
-        <Icon
+        <IconComponent
           testID="iconIcon"
           style={StyleSheet.flatten([
             { backgroundColor: 'transparent' },
@@ -76,7 +83,7 @@ Icon.propTypes = {
   name: PropTypes.string,
   size: PropTypes.number,
   color: PropTypes.string,
-  component: PropTypes.func,
+  Component: PropTypes.func,
   underlayColor: PropTypes.string,
   reverse: PropTypes.bool,
   raised: PropTypes.bool,
@@ -96,6 +103,7 @@ Icon.defaultProps = {
   color: 'black',
   reverseColor: 'white',
   disabled: false,
+  type: 'material',
 };
 
 const styles = StyleSheet.create({
@@ -104,14 +112,14 @@ const styles = StyleSheet.create({
   },
   raised: {
     ...Platform.select({
-      ios: {
+      android: {
+        elevation: 2,
+      },
+      default: {
         shadowColor: 'rgba(0,0,0, .4)',
         shadowOffset: { height: 1, width: 1 },
         shadowOpacity: 1,
         shadowRadius: 1,
-      },
-      android: {
-        elevation: 2,
       },
     }),
   },
