@@ -3,16 +3,13 @@ import PropTypes from 'prop-types';
 import {
   View,
   StyleSheet,
-  Dimensions,
   Platform,
   TouchableWithoutFeedback,
+  Modal,
 } from 'react-native';
 
 import { ViewPropTypes, withTheme } from '../config';
-
-const dimensions = Dimensions.get('window');
-const windowWidth = dimensions.width;
-const windowHeight = dimensions.height;
+import { ScreenHeight, ScreenWidth } from '../helpers';
 
 const Overlay = props => {
   const {
@@ -30,44 +27,51 @@ const Overlay = props => {
     ...rest
   } = props;
 
-  if (!isVisible) return null;
-
   return (
-    <TouchableWithoutFeedback onPress={onBackdropPress}>
-      <View
-        testID="overlayContainer"
-        style={StyleSheet.flatten([
-          styles.container,
-          { backgroundColor: windowBackgroundColor },
-          containerStyle,
-        ])}
-        {...rest}
+    <Modal
+      visible={isVisible}
+      onRequestClose={onBackdropPress}
+      transparent
+      {...rest}
+    >
+      <TouchableWithoutFeedback
+        onPress={onBackdropPress}
+        testID="RNE__Overlay__backdrop"
       >
-        <TouchableWithoutFeedback>
-          <View
-            style={StyleSheet.flatten([
-              styles.overlay,
-              {
-                borderRadius,
-                backgroundColor: overlayBackgroundColor,
-                width,
-                height,
-              },
-              fullScreen && { width: windowWidth, height: windowHeight },
-              overlayStyle,
-            ])}
-          >
-            {children}
-          </View>
-        </TouchableWithoutFeedback>
+        <View
+          testID="overlayContainer"
+          style={StyleSheet.flatten([
+            styles.backdrop,
+            { backgroundColor: windowBackgroundColor },
+            containerStyle,
+          ])}
+        />
+      </TouchableWithoutFeedback>
+
+      <View style={styles.container} pointerEvents="box-none">
+        <View
+          style={StyleSheet.flatten([
+            styles.overlay,
+            {
+              borderRadius,
+              backgroundColor: overlayBackgroundColor,
+              width,
+              height,
+            },
+            fullScreen && styles.fullscreen,
+            overlayStyle,
+          ])}
+        >
+          {children}
+        </View>
       </View>
-    </TouchableWithoutFeedback>
+    </Modal>
   );
 };
 
 Overlay.propTypes = {
-  children: PropTypes.any,
-  isVisible: PropTypes.bool,
+  children: PropTypes.element.isRequired,
+  isVisible: PropTypes.bool.isRequired,
   containerStyle: ViewPropTypes.style,
   overlayStyle: ViewPropTypes.style,
   windowBackgroundColor: PropTypes.string,
@@ -84,33 +88,41 @@ Overlay.defaultProps = {
   fullScreen: false,
   windowBackgroundColor: 'rgba(0, 0, 0, .4)',
   overlayBackgroundColor: 'white',
-  width: windowWidth - 80,
-  height: windowHeight - 180,
+  width: ScreenWidth - 80,
+  height: ScreenHeight - 180,
   onBackdropPress: () => null,
 };
 
 const styles = StyleSheet.create({
-  container: {
+  backdrop: {
     position: 'absolute',
     top: 0,
     left: 0,
-    width: windowWidth,
-    height: windowHeight,
+    width: '100%',
+    height: '100%',
     justifyContent: 'center',
     alignItems: 'center',
-    elevation: 5,
+  },
+  container: {
+    flex: 1,
+    alignItems: 'center',
+    justifyContent: 'center',
+  },
+  fullscreen: {
+    width: '100%',
+    height: '100%',
   },
   overlay: {
     borderRadius: 5,
     padding: 10,
     ...Platform.select({
-      ios: {
+      android: {
+        elevation: 2,
+      },
+      default: {
         shadowColor: 'rgba(0, 0, 0, .3)',
         shadowOffset: { width: 0, height: 1 },
         shadowRadius: 4,
-      },
-      android: {
-        elevation: 2,
       },
     }),
   },
