@@ -1,12 +1,6 @@
 import PropTypes from 'prop-types';
 import React, { Component } from 'react';
-import {
-  Dimensions,
-  StyleSheet,
-  View,
-  ActivityIndicator,
-  Text,
-} from 'react-native';
+import { StyleSheet, View, ActivityIndicator, Text } from 'react-native';
 
 import { ViewPropTypes } from '../config';
 import { nodeType, renderNode } from '../helpers';
@@ -14,28 +8,27 @@ import { nodeType, renderNode } from '../helpers';
 import Input from '../input/Input';
 import Icon from '../icons/Icon';
 
-const SCREEN_WIDTH = Dimensions.get('window').width;
 const ANDROID_GRAY = 'rgba(0, 0, 0, 0.54)';
 
 const defaultSearchIcon = {
-  type: 'material-community',
+  type: 'material',
   size: 25,
   color: ANDROID_GRAY,
-  name: 'magnify',
+  name: 'search',
 };
 
 const defaultCancelIcon = {
-  type: 'material-community',
+  type: 'material',
   size: 25,
   color: ANDROID_GRAY,
-  name: 'arrow-left',
+  name: 'arrow-back',
 };
 
 const defaultClearIcon = {
-  type: 'material-community',
-  name: 'close',
+  type: 'material',
   size: 25,
   color: ANDROID_GRAY,
+  name: 'clear',
 };
 
 class SearchBar extends Component {
@@ -75,9 +68,10 @@ class SearchBar extends Component {
 
   constructor(props) {
     super(props);
+    const { value } = props;
     this.state = {
       hasFocus: false,
-      isEmpty: true,
+      isEmpty: value ? value === '' : true,
     };
   }
 
@@ -106,7 +100,10 @@ class SearchBar extends Component {
           onFocus={this.onFocus}
           onBlur={this.onBlur}
           onChangeText={this.onChangeText}
-          ref={input => (this.input = input)}
+          ref={input => {
+            this.input = input;
+          }}
+          containerStyle={{ paddingHorizontal: 0 }}
           inputStyle={StyleSheet.flatten([styles.input, inputStyle])}
           inputContainerStyle={StyleSheet.flatten([
             styles.inputContainer,
@@ -152,6 +149,7 @@ class SearchBar extends Component {
 }
 
 SearchBar.propTypes = {
+  value: PropTypes.string,
   clearIcon: nodeType,
   searchIcon: nodeType,
   cancelIcon: nodeType,
@@ -170,6 +168,7 @@ SearchBar.propTypes = {
 };
 
 SearchBar.defaultProps = {
+  value: '',
   loadingProps: {},
   showLoading: false,
   onClear: () => null,
@@ -178,14 +177,13 @@ SearchBar.defaultProps = {
   onBlur: () => null,
   onChangeText: () => null,
   searchIcon: defaultSearchIcon,
-  clearIcon: defaultCancelIcon,
+  clearIcon: defaultClearIcon,
   cancelIcon: defaultCancelIcon,
 };
 
 const styles = StyleSheet.create({
   container: {
     backgroundColor: 'white',
-    width: SCREEN_WIDTH,
     paddingTop: 8,
     paddingBottom: 8,
   },
@@ -195,7 +193,7 @@ const styles = StyleSheet.create({
   },
   inputContainer: {
     borderBottomWidth: 0,
-    width: SCREEN_WIDTH,
+    width: '100%',
   },
   rightIconContainerStyle: {
     marginRight: 8,
