@@ -48,15 +48,14 @@ describe('Text Component', () => {
     expect(component.props().children).toBe('Children Text');
   });
 
-  it('should render fontFamily and style', () => {
+  it('should apply style prop as an array', () => {
     const component = shallow(
-      <Text fontFamily="comic-sans" style={{ color: 'red' }}>
-        Children Text
-      </Text>
+      <Text style={[{ color: 'red' }, { fontSize: 30 }]}>Children Text</Text>
     );
-
-    expect(component.length).toBe(1);
-    expect(toJson(component)).toMatchSnapshot();
+    expect(component.props().style).toEqual({
+      color: 'red',
+      fontSize: 30,
+    });
   });
 
   it('should use values from the theme', () => {
@@ -75,6 +74,28 @@ describe('Text Component', () => {
     expect(component.root.findByType(TextThemed).children[0].props.h4).toBe(
       true
     );
+    expect(component.toJSON()).toMatchSnapshot();
+  });
+
+  it('local props should override style props on theme', () => {
+    const theme = {
+      Text: {
+        style: {
+          fontSize: 14,
+        },
+      },
+    };
+
+    const component = renderer.create(
+      <ThemeProvider theme={theme}>
+        <TextThemed h2>Hey</TextThemed>
+      </ThemeProvider>
+    );
+
+    expect(
+      component.root.findByType(TextThemed).children[0].children[0].props.style
+        .fontSize
+    ).toBe(42.5);
     expect(component.toJSON()).toMatchSnapshot();
   });
 });
