@@ -12,17 +12,17 @@ const siteConfig = {
   baseUrl: '/react-native-elements/' /* base url for your project */,
   projectName: 'react-native-elements',
   organizationName: 'react-native-training',
-  defaultVersionShown: '0.19.1',
   headerLinks: [
     { doc: 'getting_started', label: 'Docs' },
     { doc: 'overview', label: 'Components' },
     { page: 'help', label: 'Help' },
     {
       href: 'https://github.com/react-native-training/react-native-elements',
-      label: 'Github',
+      label: 'GitHub',
     },
-    // { blog: true, label: 'Blog' },
+    { blog: true, label: 'Blog' },
   ],
+  onPageNav: 'separate',
   // Algolia configuration for search feature
   algolia: {
     apiKey: '89e04a9445d16350e100c2d2421f2d39',
